requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
transformers
openai
python-dotenv
